# docs/openapi.yml
openapi: 3.0.1
info:
  title: CTO
  description: All api documentation.
  version: "1.0.2"
  contact:
    name: API Support
    email: qufiwefefwoyn@gmail.com
servers:
  - url: https://ctopanel.com/api
    description: Production.
  - url: http://127.0.0.1:8888/api
    description: Local core service.
  - url: http://127.0.0.1:9010/api
    description: Local porter service.
  - url: http://127.0.0.1:8900/api
    description: Local loginput service.
...
paths:
...
  /core/temp/{projectID}:
    get:
      tags: [core]
      summary: Debug service gets key directly from badger - may be removed in future.
      parameters:
        - $ref: "#/components/parameters/projectID"
        - in: query
          name: key
          schema:
            type: string
      responses:
        "200":
          description: "OK"
  /core/setproject/{projectID}:
    post:
      tags: [core]
      summary: Set new project.
      parameters:
        - $ref: "#/components/parameters/projectID"
      responses:
        "200":
          description: "OK"
  /core/multi:
    delete:
      tags: [core]
      summary: Delete all keys with specified prefix.
      parameters:
        - in: query
          name: prefix
          schema:
            type: string
          required: true
      responses:
        "200":
          description: "OK"
...
components:
  parameters:
    projectID:
      in: path
      name: projectID
      schema:
        type: string
        format: uuid
      required: true
      description: Project ID
...
  securitySchemes:
    TokenAuth:
      type: apiKey
      name: Authorization
      in: header
      description: Token-based authentication with required prefix "Token"
    FluentdBasicAuth:
      type: http
      scheme: basic
      description: >
        Fluentd basic auth. Username is project name, password is project secret.